yup Updated ArgumentManager to support static types Updated InfiniteAmmo CustomFlag to regenerate ammo or instantly add back Readded Disguise CustomFlag for Exiled Added PlayerDamageMultiplier and DoorDamageMultiplier to ExplosiveGrenadeData Moved AudioAPI and BadgeManager into Helper folder ItemGlow now supports Intensity and range Added Events Commands now use permissions instead of RA permsissions Added Inspection ItemEvent Added Rotation to Spawn Settings Readded ECR to SwitchRoleOnUse Secret Code Cleanup

+++ UncomplicatedCustomItems/API/Features/FlagSettings/CraftableSettings.cs
@@ -7,6 +7,6 @@ namespace UncomplicatedCustomItems.API.Features
     {
         public Scp914KnobSetting? KnobSetting { get; set; } = Scp914KnobSetting.Coarse;
         public ItemType? OriginalItem { get; set; } = ItemType.Adrenaline;
-        public int? Chance { get; set; } = 100;
+        public float? Chance { get; set; } = 100;
     }
 }
