+++ UncomplicatedCustomItems/API/Interfaces/FlagSettings/ISpawnItemWhenDetonatedSettings.cs
@@ -7,7 +7,7 @@ namespace UncomplicatedCustomItems.API.Interfaces
         public abstract string? ItemType { get; set; }
         public abstract uint? ItemId { get; set; }
         public abstract float? TimeTillDespawn { get; set; }
-        public abstract uint? Chance { get; set; }
+        public abstract float? Chance { get; set; }
         public abstract bool? Pickupable { get; set; }
     }
 }
